0.1.0
Under the client folder I started with the webpack.config.js- added the html webpackplugin, workbox pugins for a service worker, manifest file, and added css loaders and babelt to webpack. Then I moved on to src-sw.js - to implement asst caching. after I went ahead and worked on the database.js to get all content from the database and accept some content and add it to database. In order to install the pwa I added event handlers to the install.js file. in the root i altered the package.json scripts to start the application and develop the dist directory